requirements.txt
langchain>=0.0.350
langchain-google-vertexai>=0.0.1
google-cloud-aiplatform>=1.38.1
python-dotenv>=1.0.0
google-generativeai>=0.3.2
fastapi>=0.105.0
uvicorn>=0.24.0
pydantic>=2.5.2
sqlalchemy>=2.0.23
pytest>=7.4.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiohttp>=3.9.1
numpy>=1.26.2
pandas>=2.1.4